a few additional updates. Getting ready to create custom api which is how the page would load data on the front end
