--- pages/catholicNewsAPI.js
@@ -1,52 +1,0 @@
-var cNewsArray = [
-{
- "title" : "After three years of ISIS occupation, the Mass returns to Mosul",
- "link" : "https://www.catholicnewsagency.com/news/after-three-years-of-isis-occupation-the-mass-returns-to-mosul-25516/" ,
- "summary" : "Aug 30, 2017 - 06:01 am .- Following the liberation of Mosul, Iraq, from the hands of the Islamic State, Christians are cautiously returning to the city. And as they return, so does the Mass.",
- "imageurl" : "https://www.catholicnewsagency.com/images/Eucharist_Credit_wideonet_Shutterstock_CNA.jpg" 
-},
-{
- "title" : "Amid downpour, Bishops consecrate Scotland to Immaculate Heart of Mary",
- "link" : "https://www.catholicnewsagency.com/news/amid-downpour-bishops-consecrate-scotland-to-immaculate-heart-of-mary-88868/" ,
- "summary" : "Sep 4, 2017 - 04:29 pm .- Constant, pouring rain did not deter the Catholic bishops of Scotland from consecrating Scotland to the Immaculate Heart of Mary on Sunday.",
- "imageurl" : "https://www.catholicnewsagency.com/images/Our_Lady_of_Fatima_Credit_Ricardo_Perna_Shutterstock_CNA.jpeg?w=640"
-},
-{
- "title" : "Amid downpour, Bishops consecrate Scotland to Immaculate Heart of Mary",
- "link" : "https://www.catholicnewsagency.com/news/amid-downpour-bishops-consecrate-scotland-to-immaculate-heart-of-mary-88868/" ,
- "summary" : "Sep 4, 2017 - 04:29 pm .- Constant, pouring rain did not deter the Catholic bishops of Scotland from consecrating Scotland to the Immaculate Heart of Mary on Sunday.",
- "imageurl" : "https://www.catholicnewsagency.com/images/Our_Lady_of_Fatima_Credit_Ricardo_Perna_Shutterstock_CNA.jpeg?w=640"
-},
-{
- "title" : "Amid downpour, Bishops consecrate Scotland to Immaculate Heart of Mary",
- "link" : "https://www.catholicnewsagency.com/news/amid-downpour-bishops-consecrate-scotland-to-immaculate-heart-of-mary-88868/" ,
- "summary" : "Sep 4, 2017 - 04:29 pm .- Constant, pouring rain did not deter the Catholic bishops of Scotland from consecrating Scotland to the Immaculate Heart of Mary on Sunday.",
- "imageurl" : "https://www.catholicnewsagency.com/images/Our_Lady_of_Fatima_Credit_Ricardo_Perna_Shutterstock_CNA.jpeg?w=640"
-},
-{
- "title" : "Amid downpour, Bishops consecrate Scotland to Immaculate Heart of Mary",
- "link" : "https://www.catholicnewsagency.com/news/amid-downpour-bishops-consecrate-scotland-to-immaculate-heart-of-mary-88868/" ,
- "summary" : "Sep 4, 2017 - 04:29 pm .- Constant, pouring rain did not deter the Catholic bishops of Scotland from consecrating Scotland to the Immaculate Heart of Mary on Sunday.",
- "imageurl" : "https://www.catholicnewsagency.com/images/Our_Lady_of_Fatima_Credit_Ricardo_Perna_Shutterstock_CNA.jpeg?w=640"
-},
-{
- "title" : "Amid downpour, Bishops consecrate Scotland to Immaculate Heart of Mary",
- "link" : "https://www.catholicnewsagency.com/news/amid-downpour-bishops-consecrate-scotland-to-immaculate-heart-of-mary-88868/" ,
- "summary" : "Sep 4, 2017 - 04:29 pm .- Constant, pouring rain did not deter the Catholic bishops of Scotland from consecrating Scotland to the Immaculate Heart of Mary on Sunday.",
- "imageurl" : "https://www.catholicnewsagency.com/images/Our_Lady_of_Fatima_Credit_Ricardo_Perna_Shutterstock_CNA.jpeg?w=640"
-},
-{
- "title" : "Amid downpour, Bishops consecrate Scotland to Immaculate Heart of Mary",
- "link" : "https://www.catholicnewsagency.com/news/amid-downpour-bishops-consecrate-scotland-to-immaculate-heart-of-mary-88868/" ,
- "summary" : "Sep 4, 2017 - 04:29 pm .- Constant, pouring rain did not deter the Catholic bishops of Scotland from consecrating Scotland to the Immaculate Heart of Mary on Sunday.",
- "imageurl" : "https://www.catholicnewsagency.com/images/Our_Lady_of_Fatima_Credit_Ricardo_Perna_Shutterstock_CNA.jpeg?w=640"
-},
-{
- "title" : "Amid downpour, Bishops consecrate Scotland to Immaculate Heart of Mary",
- "link" : "https://www.catholicnewsagency.com/news/amid-downpour-bishops-consecrate-scotland-to-immaculate-heart-of-mary-88868/" ,
- "summary" : "Sep 4, 2017 - 04:29 pm .- Constant, pouring rain did not deter the Catholic bishops of Scotland from consecrating Scotland to the Immaculate Heart of Mary on Sunday.",
- "imageurl" : "https://www.catholicnewsagency.com/images/Our_Lady_of_Fatima_Credit_Ricardo_Perna_Shutterstock_CNA.jpeg?w=640"
-}
-]
-
-export default cNewsArray
